Add tests for build and change tests in add
