+++ persist-cli-tests/src/test/resources/test-src/input/tool_test_build_7/persist/model.bal
@@ -1,0 +1,61 @@
+// Copyright (c) 2024 WSO2 LLC. (http://www.wso2.org) All Rights Reserved.
+//
+// WSO2 LLC. licenses this file to you under the Apache License,
+// Version 2.0 (the "License"); you may not use this file except
+// in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing,
+// software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied.  See the License for the
+// specific language governing permissions and limitations
+// under the License.
+
+import ballerina/time;
+import ballerina/persist as _;
+import ballerina/constraint;
+
+// Defines the entity type with the entity identity
+
+type Employee record {|
+    readonly string empNo;
+    string firstName;
+    string lastName;
+    time:Date birthDate;
+    string gender;
+    time:Date hireDate;
+
+    Department department;
+    Workspace? workspace;
+|};
+
+type Workspace record {|
+    readonly string workspaceId;
+    string workspaceType;
+
+    Building location;
+    Employee employee;
+|};
+
+type Building record {|
+    readonly string buildingCode;
+    string city;
+    string state;
+    string country;
+    string postalCode;
+
+    Workspace[] workspaces;
+|};
+
+type Department record {|
+    @constraint:String {
+        maxLength: 10
+    }
+    readonly string deptNo;
+    string deptName;
+
+    Employee[] employees;
+|};
