+++ persist-cli-tests/src/test/resources/test-src/output/tool_test_add_1/Ballerina.toml
@@ -17,7 +17,7 @@ array = [ [1, 0], [0, 1] ]
 [[platform.java17.dependency]]
 groupId = "io.ballerina.stdlib"
 artifactId = "persist.sql-native"
-version = "1.3.0-SNAPSHOT"
+version = "1.4.0-SNAPSHOT"
 
 [[tool.persist]]
 id = "generate-db-client"
